Only expose the minimal API.
