+++ src/util.ts
@@ -1,4 +1,15 @@
+import CID from 'cids';
+import multihashing from 'multihashing-async';
 
+export enum Action {
+    get = 'GET',
+    put = 'PUT',
+    delete = 'DEL',
+    list = 'LIST'
+}
+
+export const makeCid = async (content: Uint8Array): Promise<string> => new CID(1, 'raw', await multihashing(content, 'blake2b-256')).toString('base58btc')
+export const makeCidString = async (content: string): Promise<string> => await makeCid(new TextEncoder().encode(content))
 
 export const makeKRI = (suffix: string, name: string, app?: string, path?: string, fragment?: string) =>
     `kepler:${suffix}://${name}${app ? '/' + app : ''}${path ? '/' + path : ''}${fragment ? '#' + fragment : ''}`
